.requirements.txt
openai
opencv-contrib-python
python-dotenv
torch
torchvision
numpy
tqdm
imutils
matplotlib
ultralytics
